Rename and move class ProgressUpdator to main
